Added scheduling stuff, moved other stuff to libIOStream and libZUtil.

+++ src/game/Coordinate.hpp
@@ -6,7 +6,8 @@
 
 namespace MCServer {
 
-typedef int32_t Coordinate;
+//typedef int32_t Coordinate;
+typedef double Coordinate;
 
 }
